pip-requirements.txt
accelerate==1.8.1
addict==2.4.0
antlr4-python3-runtime==4.9.3
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
boto3==1.39.6
botocore==1.39.6
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
contourpy==1.3.2
controlnet_aux==0.0.7
cupy-cuda12x==13.5.1
cycler==0.12.1
dash==3.1.1
decorator==5.2.1
diff_surfel_rasterization @ file:///hpc2hdd/home/hongfeizhang/hongfei_workspace/DiffSynth-Studio/GenFusion/Reconstruction/submodules/diff-surfel-rasterization
-e git+ssh://git@github.com/SarahAnberlin/DiffSynth-Studio.git@0cc6fadfc91a798418b8903217f2c68c2b4cbb4c#egg=diffsynth
diffusers==0.34.0
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.5
fsspec==2025.5.1
ftfy==6.3.1
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
minio==7.2.15
modelscope==1.28.0
mpmath==1.3.0
narwhals==1.48.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
open3d==0.19.0
opencv-contrib-python==4.12.0.88
opencv-python==4.10.0.84
packaging==25.0
pandas==2.3.1
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
plyfile==1.1.2
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycolmap @ git+https://github.com/rmbrualla/pycolmap@cc7ea4b7301720ac29287dbe450952511b32125e
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pynvml==12.0.0
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytransform3d==3.14.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
retrying==1.4.1
rpds-py==0.26.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
sentencepiece==0.2.0
simple_knn @ file:///hpc2hdd/home/hongfeizhang/hongfei_workspace/DiffSynth-Studio/GenFusion/Reconstruction/submodules/simple-knn
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tartanair==1.1.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.17
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
triton==3.3.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.23.0
